+++ nuxt/base/.playground/app.vue
@@ -1,0 +1,7 @@
+<template>
+  <div>
+		Hello there! 👋
+
+		{{ useAppConfig().base.name }}
+	</div>
+</template>